fix(vacation.entity): type 수정
- proxyMemberId 타입을 LONG으로 수정했습니다

+++ src/main/java/kr/co/hr/vacation/entity/Vacation.java
@@ -35,7 +35,7 @@ public class Vacation {
     private LocalDate startDate; 
     private LocalDate endDate;   
     private Integer days;        
-    private Integer proxyMemberId;
+    private Long proxyMemberId;
     
     @PrePersist
     public void prePersist() {
